# docs/OPENAPI.JSON
{
  "openapi": "3.1.0",
  "info": {
    "title": "Savory Canvas API",
    "version": "1.2.1",
    "description": "Savory Canvas 本地图文生成系统接口文档"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8887",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
    "/api/v1/sessions": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "创建会话",
        "operationId": "createSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "会话列表",
        "operationId": "listSessions",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "会话详情",
        "operationId": "getSessionDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "patch": {
        "tags": [
          "Session"
        ],
        "summary": "重命名会话",
        "operationId": "updateSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Session"
        ],
        "summary": "删除会话",
        "operationId": "deleteSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/styles/chat": {
      "post": {
        "tags": [
          "Style"
        ],
        "summary": "风格对话",
        "operationId": "styleChat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "返回选项",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleChatResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/styles": {
      "post": {
        "tags": [
          "Style"
        ],
        "summary": "保存风格模板",
        "operationId": "createStyleProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleProfileCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfile"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "get": {
        "tags": [
          "Style"
        ],
        "summary": "风格列表",
        "operationId": "listStyleProfiles",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfileListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/styles/{style_id}": {
      "patch": {
        "tags": [
          "Style"
        ],
        "summary": "更新风格",
        "operationId": "updateStyleProfile",
        "parameters": [
          {
            "$ref": "#/components/parameters/StyleId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleProfile"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Style"
        ],
        "summary": "删除风格",
        "operationId": "deleteStyleProfile",
        "parameters": [
          {
            "$ref": "#/components/parameters/StyleId"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/providers": {
      "get": {
        "tags": [
          "Provider"
        ],
        "summary": "提供商列表",
        "operationId": "listProviders",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Provider"
        ],
        "summary": "新增提供商",
        "operationId": "createProvider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/providers/{provider_id}": {
      "patch": {
        "tags": [
          "Provider"
        ],
        "summary": "更新提供商",
        "operationId": "updateProvider",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Provider"
        ],
        "summary": "删除提供商",
        "operationId": "deleteProvider",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderId"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "SessionId": {
        "name": "session_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "StyleId": {
        "name": "style_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "ProviderId": {
        "name": "provider_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "错误响应",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "code",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "SessionCreateRequest": {
        "type": "object",
        "required": [
          "title",
          "content_mode"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content_mode": {
            "type": "string",
            "enum": [
              "food",
              "scenic",
              "food_scenic"
            ]
          }
        }
      },
      "SessionUpdateRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "Session": {
        "type": "object",
        "required": [
          "id",
          "title",
          "content_mode",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content_mode": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SessionListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          }
        }
      },
      "SessionDetailResponse": {
        "type": "object",
        "required": [
          "session",
          "assets",
          "jobs",
          "exports"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/Session"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationJob"
            }
          },
          "exports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExportTask"
            }
          }
        }
      },
...
      "Asset": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "asset_type",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "asset_type": {
            "type": "string",
            "enum": [
              "food_name",
              "scenic_name",
              "text",
              "image",
              "video",
              "transcript"
            ]
          },
          "content": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "processing",
              "failed"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StyleChatRequest": {
        "type": "object",
        "required": [
          "session_id",
          "stage",
          "user_reply"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "stage": {
            "type": "string",
            "enum": [
              "init",
              "painting_style",
              "background_decor",
              "color_mood",
              "image_count"
            ]
          },
          "user_reply": {
            "type": "string"
          },
          "selected_items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StyleOptionBlock": {
        "type": "object",
        "required": [
          "title",
          "items",
          "max"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "StyleChatResponse": {
        "type": "object",
        "required": [
          "reply",
          "options",
          "stage",
          "next_stage",
          "is_finished",
          "fallback_used"
        ],
        "properties": {
          "reply": {
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/StyleOptionBlock"
          },
          "stage": {
            "type": "string"
          },
          "next_stage": {
            "type": "string"
          },
          "is_finished": {
            "type": "boolean"
          },
          "fallback_used": {
            "type": "boolean"
          }
        }
      },
      "StylePayload": {
        "type": "object",
        "required": [
          "painting_style",
          "color_mood",
          "prompt_example",
          "style_prompt",
          "extra_keywords"
        ],
        "properties": {
          "painting_style": {
            "type": "string"
          },
          "color_mood": {
            "type": "string"
          },
          "prompt_example": {
            "type": "string"
          },
          "style_prompt": {
            "type": "string"
          },
          "sample_image_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "extra_keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StyleProfileCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "style_payload"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "style_payload": {
            "$ref": "#/components/schemas/StylePayload"
          }
        }
      },
      "StyleProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "style_payload": {
            "$ref": "#/components/schemas/StylePayload"
          }
        }
      },
      "StyleProfile": {
        "type": "object",
        "required": [
          "id",
          "name",
          "style_payload",
          "is_builtin",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "style_payload": {
            "$ref": "#/components/schemas/StylePayload"
          },
          "sample_image_preview_url": {
            "type": "string",
            "format": "uri"
          },
          "is_builtin": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StyleProfileListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StyleProfile"
            }
          }
        }
      },
...
      "GenerationJob": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "status",
          "progress_percent",
          "current_stage",
          "stage_message",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "style_profile_id": {
            "type": "string"
          },
          "image_count": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "partial_success",
              "success",
              "failed",
              "canceled"
            ]
          },
          "progress_percent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "current_stage": {
            "type": "string",
            "enum": [
              "asset_extract",
              "asset_allocate",
              "prompt_generate",
              "image_generate",
              "copy_generate",
              "finalize"
            ]
          },
          "stage_message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ExportTask": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "job_id",
          "export_format",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          },
          "export_format": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "success",
              "failed"
            ]
          },
          "file_url": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProviderCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "base_url",
          "api_key",
          "api_protocol"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "api_key": {
            "type": "string"
          },
          "api_protocol": {
            "type": "string",
            "enum": [
              "responses",
              "chat_completions"
            ]
          }
        }
      },
      "ProviderUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "api_key": {
            "type": "string"
          },
          "api_protocol": {
            "type": "string",
            "enum": [
              "responses",
              "chat_completions"
            ]
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "Provider": {
        "type": "object",
        "required": [
          "id",
          "name",
          "base_url",
          "api_protocol",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "api_key_masked": {
            "type": "string"
          },
          "api_protocol": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProviderListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Provider"
            }
          }
        }
      },
...
      "DeleteResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}